fix : 머지하면서 생긴 문제들 갈아엎기
- keeper_files git ignore 처리
- BookEntity에 picture 컬럼 제거
- Id 컬럼 전부 Int -> Long 변경

+++ src/main/java/keeper/project/homepage/controller/member/MemberController.java
@@ -37,7 +37,7 @@ public class MemberController {
   public SingleResult<MemberEntity> findMember() {
     // SecurityContext에서 인증받은 회원의 정보를 얻어온다.
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    Integer id = Integer.valueOf(authentication.getName());
+    Long id = Long.valueOf(authentication.getName());
     // 결과데이터가 단일건인경우 getSingleResult를 이용해서 결과를 출력한다.
     return responseService.getSingleResult(
         memberRepository.findById(id).orElseThrow(CustomMemberNotFoundException::new));
