+++ src/main/java/keeper/project/homepage/repository/member/MemberTypeRepository.java
@@ -3,6 +3,6 @@ package keeper.project.homepage.repository.member;
 import keeper.project.homepage.entity.member.MemberTypeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberTypeRepository extends JpaRepository<MemberTypeEntity, Integer> {
+public interface MemberTypeRepository extends JpaRepository<MemberTypeEntity, Long> {
 
 }
